Pov #3 (#12)
* Cleaned up constants

* Working angle alignment

* Removed FUNDING.yml so repo uses the one in the .github repo

* Moved some stuff around

* Update to Advantage Scope Beta 3

+++ src/main/java/com/koibots/robot/Constants.java
@@ -8,7 +8,8 @@ import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj.smartdashboard.Field2d;
 
 public class Constants {
-  public static final double MAX_LINEAR_SPEED_METERS_PER_SECOND = 4; // Meters per Second
+
+  public static final int MAX_LINEAR_SPEED_METERS_PER_SECOND = 4; // Meters per Second
   public static final double MAX_ANGULAR_VELOCITY_RADIANS_PER_SECOND = 2 * PI; // Radians per Second
 
   public static final double DEADBAND = 0.02;
